Moved Levenshtein Distrance algorithm into separate trait
